refactor: change name file for constant values
